working better

+++ project02/src/include/Golomb.hh
@@ -2,18 +2,18 @@
 #define GOLOMB_HH
 
 #include "./BitStream.hh"
-#include <sys/types.h>
+#include <cstdint>
 
 class Golomb {
 private:
-  uint m;
-  uint b;
-  uint c;
+  uint32_t m;
+  uint32_t b;
+  uint32_t c;
   BitStream *bs;
 public:
-  Golomb(uint m, BitStream *bs);
-  int encode(int n);
-  int decode(int *g);
+  Golomb(uint32_t m, BitStream *bs);
+  int64_t encode(int64_t n);
+  int64_t decode(int64_t *g);
 };
 
 #endif
